+++ changelog.md
@@ -14,6 +14,8 @@
 - Updated GitHub Actions CI config to run tests with Node.js v12, v14, v16.
 - Simplified JSDoc related package scripts now that [`jsdoc-md`](https://npm.im/jsdoc-md) v10+ automatically generates a Prettier formatted readme.
 - Added a package `test:jsdoc` script that checks the readme API docs are up to date with the source JSDoc.
+- Test the bundle size using [`esbuild`](https://npm.im/esbuild) instead of [`webpack`](https://npm.im/webpack) and [`disposable-directory`](https://npm.im/disposable-directory).
+- Increased the documented bundle size to “< 4 kB” to match that of [`esbuild`](https://npm.im/esbuild) instead of [`webpack`](https://npm.im/webpack).
 - Use the correct `kB` symbol instead of `KB` wherever bundle size is mentioned in the package description and readme.
 - Updated the [example Next.js app](https://graphql-react.vercel.app) URL in the readme.